Modernized build scripts

+++ Nuclex.Pixels.Native/build.sh
@@ -10,23 +10,53 @@ for buildMode in Debug Release; do
 	# enter each and every source file into your CMakeLists.txt. See the
 	# documentation on file(GLOB_RECURSE ...) for that turd.
 	#
-	if [ -d obj ]; then
-		if [ -d obj/cmake-$buildMode ]; then
-			rm -rf obj/cmake-$buildMode
+	if [ $# -ge 1 ] && [ $1 == "full" ]; then
+		if [ -d obj ]; then
+			if [ -d obj/cmake-$buildMode ]; then
+				rm -rf obj/cmake-$buildMode
+			fi
 		fi
 	fi
 
 	# Let CMake build a Makefile or a Ninja build script
 	if [ $# -ge 1 ] && [ $1 == "ninja" ]; then
-		cmake -B obj/cmake-$buildMode -DCMAKE_BUILD_TYPE=$buildMode -GNinja
+		cmake \
+			-B obj/cmake-$buildMode \
+			-D CMAKE_BUILD_TYPE=$buildMode \
+			-D WANT_JPEGXL=OFF \
+			-D WANT_AVIF=OFF \
+			-D WANT_WEBP=ON \
+			-D WANT_TIFF=ON \
+			-D WANT_JPG=ON \
+			-D WANT_EXR=ON \
+			-D WANT_PNG=ON \
+			-D BUILD_UNIT_TESTS=ON \
+			-D BUILD_BENCHMARK=ON \
+			-GNinja
 	else
-		cmake -B obj/cmake-$buildMode -DCMAKE_BUILD_TYPE=$buildMode
+		cmake \
+			-B obj/cmake-$buildMode \
+			-D CMAKE_BUILD_TYPE=$buildMode \
+			-D WANT_JPEGXL=OFF \
+			-D WANT_AVIF=OFF \
+			-D WANT_WEBP=ON \
+			-D WANT_TIFF=ON \
+			-D WANT_JPG=ON \
+			-D WANT_EXR=ON \
+			-D WANT_PNG=ON \
+			-D BUILD_UNIT_TESTS=ON \
+			-D BUILD_BENCHMARK=ON
 	fi
 
 	# Compile the binary
-	cmake --build obj/cmake-$buildMode --config $buildMode --parallel $processorCount
+	cmake \
+		--build obj/cmake-$buildMode \
+		--config $buildMode \
+		--parallel $processorCount
 
 	# Put build artifacts in ./bin/linux-gcc9.2-amd64-release/ or similar
-	cmake --install obj/cmake-$buildMode --config $buildMode
+	cmake \
+		--install obj/cmake-$buildMode \
+		--config $buildMode
 
 done
